Added additional methods to play the game with
